Moved glue.qt.qtutil.RGBEdit to glue.viewers.image.qt.rgb_edit
