+++ CHANGES.md
@@ -92,6 +92,7 @@ v0.7 (unreleased)
   ``glue.qt.qtutil.data_wizard``                     | ``glue.dialogs.data_wizard.qt.data_wizard``
   ``glue.qt.mime``                                   | ``glue.core.qt.mime``
   ``glue.qt.qtutil.ComponentIDCombo``                | ``glue.core.qt.component_id_combo``
+  ``glue.qt.qtutil.RGBEdit``                         | ``glue.viewers.image.qt.rgb_edit.RGBEdit``
 
 * Improved under-the-hood creation of ROIs for Scatter and Histogram Clients. [#676]
